+++ docs/scripts/util/checkLink.ts
@@ -14,6 +14,7 @@ export type LinkInfo = {
   statusCode: number;
   tagName: string;
   tagText: string;
+  originalStatusCode?: number;
 };
 const requestedUrl: Set<string> = new Set();
 
@@ -39,9 +40,11 @@ async function returnStatus({
       );
     }
     /**
-     * If 308, check if it's a internal direction (see docs/next.config.js redirects logic)
-     * If it's internal direction, after adding the platform, it should be 200
-     * Otherwise, the link needs to be updated
+     * If 308, check if it's an internal redirection to add [platform] prefix. See docs/next.config.js redirects logic.
+     * If it's an internal redirection to add [platform] prefix, after adding the platform, it should be 200.
+     * Otherwise, the link needs to be updated.
+     * e.g. 308 from "/guides" passes (because it's redirected to "react/guides"),
+     *  but 308 from "connected-components/authenticator/headless" fails and returns a 404 ((because it's redirected to "react/connected-components/authenticator/advanced").
      */
     if (statusCode === 308) {
       const hostNameRegex = RegExp(`http(s)?:\/\/[^/]*`, 'i'); // matches everything between http(s)?: to "/", which is the hostname. e.g., "https://github.com/".
@@ -94,6 +97,7 @@ export async function checkLink(
     } else {
       const { get } = href.includes('https:') ? https : http;
       const request = await get(href, async ({ statusCode = 0 }) => {
+        const originalStatusCode = statusCode;
         statusCode =
           (
             await returnStatus({
@@ -102,7 +106,7 @@ export async function checkLink(
             })
           )?.statusCode ?? statusCode;
         requestedUrl.add(href);
-        res({ ...linkData, statusCode });
+        res({ ...linkData, statusCode, originalStatusCode });
       });
       request.end();
     }