+++ packages/ui/src/theme/tokens/components/field.ts
@@ -6,16 +6,18 @@ type FieldSizeTokens<Output> = DesignTokenProperties<
 >;
 
 export type FieldTokens<Output extends OutputVariantKey> =
-  FieldSizeTokens<Output> & {
-    small?: FieldSizeTokens<Output>;
-    large?: FieldSizeTokens<Output>;
-    label?: DesignTokenProperties<'color', Output>;
-  };
+  FieldSizeTokens<Output> &
+    DesignTokenProperties<'flexDirection', Output> & {
+      small?: FieldSizeTokens<Output>;
+      large?: FieldSizeTokens<Output>;
+      label?: DesignTokenProperties<'color', Output>;
+    };
 
 export const field: Required<FieldTokens<'default'>> = {
   // default styles
   gap: { value: '{space.xs.value}' },
   fontSize: { value: '{fontSizes.medium.value}' },
+  flexDirection: { value: 'column' },
 
   // Adjust base fontSize and gap for small and large sizes
   small: {